rename utilities/hypre_*.h
