+++ src/matrix_matrix/HYPRE_ConvertParCSRMatrixToDistributedMatrix.c
@@ -17,7 +17,7 @@
 
 #include <HYPRE_config.h>
 
-#include "hypre_general.h"
+#include "general.h"
 
 #include "_hypre_utilities.h"
 #include "HYPRE.h"
