+++ src/parcsr_ls/par_strength.c
@@ -13,7 +13,7 @@
 
 
 #include "_hypre_parcsr_ls.h"
-#include "hypre_hopscotch_hash.h"
+#include "hopscotch_hash.h"
 
 
 /*==========================================================================*/
