+++ src/parcsr_mv/par_csr_matop.c
@@ -6,7 +6,7 @@
  ******************************************************************************/
 
 #include "_hypre_utilities.h"
-#include "hypre_hopscotch_hash.h"
+#include "hopscotch_hash.h"
 #include "_hypre_parcsr_mv.h"
 #include "_hypre_lapack.h"
 #include "_hypre_blas.h"
