+++ src/utilities/CMakeLists.txt
@@ -6,7 +6,7 @@
 set(HDRS
   HYPRE_utilities.h
   _hypre_utilities.h
-  hypre_hopscotch_hash.h
+  hopscotch_hash.h
   HYPRE_error_f.h
   fortran.h
   fortran_matrix.h
@@ -21,22 +21,22 @@ set(SRCS
   F90_HYPRE_error.c
   F90_HYPRE_general.c
   fortran_matrix.c
-  hypre_ap.c
-  hypre_complex.c
-  hypre_cuda_utils.c
-  hypre_error.c
-  hypre_general.c
+  ap.c
+  complex.c
+  cuda_utils.c
+  error.c
+  general.c
   handle.c
-  hypre_hopscotch_hash.c
-  hypre_memory.c
-  hypre_merge_sort.c
-  hypre_mpi_comm_f2c.c
-  hypre_nvtx.c
-  hypre_omp_device.c
-  hypre_prefix_sum.c
-  hypre_printf.c
-  hypre_qsort.c
-  hypre_utilities.c
+  hopscotch_hash.c
+  memory.c
+  merge_sort.c
+  mpi_comm_f2c.c
+  nvtx.c
+  omp_device.c
+  prefix_sum.c
+  printf.c
+  qsort.c
+  utilities.c
   mpistubs.c
   qsplit.c
   random.c
@@ -53,12 +53,12 @@ target_sources(${PROJECT_NAME}
 if (HYPRE_USING_CUDA)
   set(CUDA_SRCS
     HYPRE_handle.c
-    hypre_cuda_utils.c
+    cuda_utils.c
     handle.c
-    hypre_general.c
-    hypre_memory.c
-    hypre_omp_device.c
-    hypre_nvtx.c
+    general.c
+    memory.c
+    omp_device.c
+    nvtx.c
   )
   convert_filenames_to_full_paths(CUDA_SRCS)
   set(HYPRE_CUDA_SOURCES ${HYPRE_CUDA_SOURCES} ${CUDA_SRCS} PARENT_SCOPE)
