+++ src/utilities/_hypre_utilities.h
@@ -77,7 +77,7 @@ typedef double                 hypre_double;
 
 #include <stdio.h>
 
-/* hypre_printf.c */
+/* printf.c */
 // #ifdef HYPRE_BIGINT
 HYPRE_Int hypre_ndigits( HYPRE_BigInt number );
 HYPRE_Int hypre_printf( const char *format , ... );
@@ -726,7 +726,7 @@ typedef struct
  * Prototypes
  *--------------------------------------------------------------------------*/
 
-/* hypre_memory.c */
+/* memory.c */
 void * hypre_Memset(void *ptr, HYPRE_Int value, size_t num, HYPRE_MemoryLocation location);
 void   hypre_MemPrefetch(void *ptr, size_t size, HYPRE_MemoryLocation location);
 void * hypre_MAlloc(size_t size, HYPRE_MemoryLocation location);
@@ -1373,7 +1373,7 @@ HYPRE_Int hypre_BinarySearch2 ( HYPRE_Int *list , HYPRE_Int value , HYPRE_Int lo
 HYPRE_Int *hypre_LowerBound( HYPRE_Int *first, HYPRE_Int *last, HYPRE_Int value );
 HYPRE_BigInt *hypre_BigLowerBound( HYPRE_BigInt *first, HYPRE_BigInt *last, HYPRE_BigInt value );
 
-/* hypre_complex.c */
+/* complex.c */
 #ifdef HYPRE_COMPLEX
 HYPRE_Complex hypre_conj( HYPRE_Complex value );
 HYPRE_Real    hypre_cabs( HYPRE_Complex value );
@@ -1386,7 +1386,7 @@ HYPRE_Real    hypre_cimag( HYPRE_Complex value );
 #define hypre_cimag(value) 0.0
 #endif
 
-/* hypre_general.c */
+/* general.c */
 #ifdef HYPRE_USING_MEMORY_TRACKER
 hypre_MemoryTracker* hypre_memory_tracker();
 #endif
@@ -1400,7 +1400,7 @@ HYPRE_Int hypre_GetDeviceLastError();
 HYPRE_Int hypre_UmpireInit(hypre_Handle *hypre_handle_);
 HYPRE_Int hypre_UmpireFinalize(hypre_Handle *hypre_handle_);
 
-/* hypre_qsort.c */
+/* qsort.c */
 void hypre_swap ( HYPRE_Int *v , HYPRE_Int i , HYPRE_Int j );
 void hypre_swap_c ( HYPRE_Complex *v , HYPRE_Int i , HYPRE_Int j );
 void hypre_swap2 ( HYPRE_Int *v , HYPRE_Real *w , HYPRE_Int i , HYPRE_Int j );
@@ -1444,7 +1444,7 @@ HYPRE_Int hypre_DoubleQuickSplit ( HYPRE_Real *values , HYPRE_Int *indices , HYP
 /* HYPRE_CUDA_GLOBAL */ HYPRE_Int hypre_RandI ( void );
 /* HYPRE_CUDA_GLOBAL */ HYPRE_Real hypre_Rand ( void );
 
-/* hypre_prefix_sum.c */
+/* prefix_sum.c */
 /**
  * Assumed to be called within an omp region.
  * Let x_i be the input of ith thread.
@@ -1481,7 +1481,7 @@ void hypre_prefix_sum_triple(HYPRE_Int *in_out1, HYPRE_Int *sum1, HYPRE_Int *in_
  */
 void hypre_prefix_sum_multiple(HYPRE_Int *in_out, HYPRE_Int *sum, HYPRE_Int n, HYPRE_Int *workspace);
 
-/* hypre_hopscotch_hash.c */
+/* hopscotch_hash.c */
 
 #ifdef HYPRE_USING_OPENMP
 
@@ -1589,7 +1589,7 @@ typedef struct
    hypre_BigHopscotchBucket* volatile table;
 } hypre_UnorderedBigIntMap;
 
-/* hypre_merge_sort.c */
+/* merge_sort.c */
 /**
  * Why merge sort?
  * 1) Merge sort can take advantage of eliminating duplicates.
@@ -1617,12 +1617,12 @@ HYPRE_Int hypre_CurandUniformSingle( HYPRE_Int n, float *urand, HYPRE_Int set_se
 
 HYPRE_Int hypre_bind_device(HYPRE_Int myid, HYPRE_Int nproc, MPI_Comm comm);
 
-/* hypre_nvtx.c */
+/* nvtx.c */
 void hypre_GpuProfilingPushRangeColor(const char *name, HYPRE_Int cid);
 void hypre_GpuProfilingPushRange(const char *name);
 void hypre_GpuProfilingPopRange();
 
-/* hypre_utilities.c */
+/* utilities.c */
 HYPRE_Int hypre_multmod(HYPRE_Int a, HYPRE_Int b, HYPRE_Int mod);
 void hypre_partition1D(HYPRE_Int n, HYPRE_Int p, HYPRE_Int j, HYPRE_Int *s, HYPRE_Int *e);
 char *hypre_strcpy(char *destination, const char *source);
