+++ src/utilities/general.h
