+++ src/utilities/headers
@@ -33,19 +33,19 @@ extern "C" {
 # Structures and prototypes
 #===========================================================================
 
-cat hypre_general.h            >> $INTERNAL_HEADER
-cat hypre_printf.h             >> $INTERNAL_HEADER
-cat hypre_error.h              >> $INTERNAL_HEADER
+cat general.h                  >> $INTERNAL_HEADER
+cat printf.h                   >> $INTERNAL_HEADER
+cat error.h                    >> $INTERNAL_HEADER
 cat mpistubs.h                 >> $INTERNAL_HEADER
-cat hypre_smp.h                >> $INTERNAL_HEADER
-cat hypre_memory.h             >> $INTERNAL_HEADER
-cat hypre_omp_device.h         >> $INTERNAL_HEADER
+cat smp.h                      >> $INTERNAL_HEADER
+cat memory.h                   >> $INTERNAL_HEADER
+cat omp_device.h               >> $INTERNAL_HEADER
 cat threading.h                >> $INTERNAL_HEADER
 cat timing.h                   >> $INTERNAL_HEADER
 cat amg_linklist.h             >> $INTERNAL_HEADER
 cat exchange_data.h            >> $INTERNAL_HEADER
 cat caliper_instrumentation.h  >> $INTERNAL_HEADER
-cat hypre_handle.h             >> $INTERNAL_HEADER
+cat handle.h                   >> $INTERNAL_HEADER
 cat gselim.h                   >> $INTERNAL_HEADER
 cat protos.h                   >> $INTERNAL_HEADER
 
@@ -87,10 +87,10 @@ extern "C++" {
 # Structures and prototypes
 #===========================================================================
 
-cat hypre_umpire_allocator.h     >> $INTERNAL_HEADER
-cat hypre_cuda_utils.h           >> $INTERNAL_HEADER
-cat hypre_cuda_reducer.h         >> $INTERNAL_HEADER
-cat hypre_cub_allocator.h        >> $INTERNAL_HEADER
+cat umpire_allocator.h     >> $INTERNAL_HEADER
+cat cuda_utils.h           >> $INTERNAL_HEADER
+cat cuda_reducer.h         >> $INTERNAL_HEADER
+cat cub_allocator.h        >> $INTERNAL_HEADER
 
 #===========================================================================
 # Include guards
