+++ src/utilities/memory.c
