+++ src/utilities/merge_sort.c
@@ -6,7 +6,7 @@
  ******************************************************************************/
 
 #include "_hypre_utilities.h"
-#include "hypre_hopscotch_hash.h"
+#include "hopscotch_hash.h"
 #include "../seq_mv/HYPRE_seq_mv.h"
 //#define DBG_MERGE_SORT
 #ifdef DBG_MERGE_SORT
