+++ src/utilities/omp_device.c
