+++ src/utilities/smp.h
